Add files via upload
This project is laundry mangement project. Where user can book a slot for picking up package and after the finishing of the work the package will be delievered to user. this also includes a admin panel where owner can handel everything. It also includes to send a request to the admin to open a new branch with same name. This has differnt panel for management.